完成加好友，加群的基本处理逻辑

+++ src/main/java/com/zps/imchat/bean/MyFriends.java
@@ -12,13 +12,13 @@ public class MyFriends {
 
     private String nickname;
 
-    private Long statu;
+    private Integer statu;
 
-    public Long getStatu() {
+    public Integer getStatu() {
         return statu;
     }
 
-    public void setStatu(Long statu) {
+    public void setStatu(Integer statu) {
         this.statu = statu;
     }
